+++ cls_hybrid_call.py
@@ -281,7 +281,7 @@ class NeuralNet:
         pred_classes = np.array([np.argmax(x)+1 for x in call_predictions_bat])[:, np.newaxis]
         
         # perform non max suppression
-        pos, prob, pred_classes, call_predictions_not_bat, features_classif = nms.nms_1d(high_preds[:,0].astype(np.float), pred_classes, 
+        pos, prob, pred_classes, call_predictions_not_bat, features_classif = nms.nms_1d(high_preds[:,0].astype(np.float32), pred_classes, 
                                                                     call_predictions_not_bat, features_classif, self.params.nms_win_size, file_duration)
   
         # remove pred that have a higher probability of not being a bat
