+++ cls_hybrid_cnn.py
@@ -277,7 +277,7 @@ class NeuralNet:
         for i in range(1,8):
             call_predictions_bat = y_predictions[:,i]
             pred_classes = np.repeat(i, call_predictions_bat.shape[0])
-            pos, prob, pred_classes, call_predictions_not_bat_nms = nms.nms_1d(call_predictions_bat.astype(np.float),
+            pos, prob, pred_classes, call_predictions_not_bat_nms = nms.nms_1d(call_predictions_bat.astype(np.float32),
                                     pred_classes, call_predictions_not_bat, self.params.nms_win_size, file_duration)
 
             # remove pred that have a higher probability of not being a bat
